[#54] makes app-container smart to add error message capabilities. adds new dumb component "components/app.jsx".  renames root app.jsx --> index.jsx (to avoid confusion).
