--- src/js/components/app-container.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-class AppContainer extends React.Component {
-
-    render () {
-
-        return (
-            <div>
-                {this.props.children}
-            </div>
-        );
-    }
-}
-
-export default AppContainer;
