+++ src/js/components/app.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const App = ({ children, error }) => {
+
+    return (
+        <div>
+            {
+                error &&
+                <h3>Error {error.status} {error.data.error}: {error.data.message}</h3>
+            }
+            { children }
+        </div>
+    );
+};
+
+export default App;
