+++ webpack.development.config.js
@@ -3,7 +3,7 @@ const webpack = require('webpack');
 
 module.exports = {
     entry: [
-        "./src/js/app.jsx",
+        "./src/js/index.jsx",
         "webpack/hot/dev-server",
         "webpack-dev-server/client?http://localhost:8080/"
     ],
